requirements.txt
# LocalChat - 100% Offline AI Document Assistant
# Python 3.7+  |  Ollama required (ollama.ai)
# Install: pip install --user -r requirements.txt

# Web framework
flask>=3.0.0
flask-cors>=4.0.0

# Production WSGI (Windows-safe)
waitress>=3.0.0

# LangChain (text splitting + FAISS wrapper)
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# Vector store
faiss-cpu>=1.7.4

# PDF parsing
pdfplumber>=0.11.0

# Image OCR
pytesseract>=0.3.10
Pillow>=10.0.0

# HTTP client (Ollama API calls)
requests>=2.32.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.26.0

# Multi-provider AI clients
openai>=1.30.0
google-generativeai>=0.7.0
cohere>=5.0.0
